+++ lib/rubocop/cop/rails_deprecation_cops.rb
@@ -1,1 +1,4 @@
 # frozen_string_literal: true
+
+require_relative 'rails_deprecation/base'
+require_relative 'rails_deprecation/to_formatted_s'