+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ### Added
 
+- Add RailsDeprecation/ToFormattedS cop.
 - Initial release.